+++ e2e/tests/plugins.e2e.js
@@ -122,6 +122,7 @@ describe('plugins', () => {
         .get(`[data-hook=${'image-item'}]`)
         .eq(1);
       cy.openPluginToolbar(PLUGIN_COMPONENT.GALLERY).shrinkPlugin();
+      cy.waitForDocumentMutations();
       cy.eyesCheckWindow(this.test.title + ' toolbar');
       cy.openGalleryAdvancedSettings();
       cy.eyesCheckWindow(this.test.title + ' settings');
@@ -131,6 +132,7 @@ describe('plugins', () => {
       cy.loadEditorAndViewer('gallery-out-of-view');
       cy.eyesCheckWindow(`${this.test.title} - out of view`);
       cy.scrollTo('bottom');
+      cy.waitForDocumentMutations();
       cy.eyesCheckWindow(`${this.test.title} - in view`);
     });
 
@@ -481,7 +483,6 @@ describe('plugins', () => {
       cy.eyesCheckWindow(this.test.title);
     });
     it('delete link preview', function() {
-      cy.moveCursorToStart();
       cy.openPluginToolbar(PLUGIN_COMPONENT.LINK_PREVIEW).wait(100);
       cy.clickToolbarButton('blockButton_delete');
       cy.triggerLinkPreviewViewerUpdate();
@@ -503,6 +504,34 @@ describe('plugins', () => {
 
     it('should embed link that supports embed', function() {
       cy.insertLinkAndEnter('www.mockUrl.com');
+      cy.eyesCheckWindow(this.test.title);
+    });
+  });
+
+  context('list', () => {
+    before(function() {
+      eyesOpen(this);
+    });
+
+    beforeEach('load editor', () => cy.loadEditorAndViewer());
+
+    after(() => cy.eyesClose());
+    it('create nested lists using tab & shift-tab', function() {
+      cy.loadEditorAndViewer()
+        .enterParagraphs(['1. Hey I am an ordered list in depth 1.'])
+        .tab()
+        .enterParagraphs(['\n Hey I am an ordered list in depth 2.'])
+        .tab()
+        .enterParagraphs(['\n Hey I am an ordered list in depth 1.'])
+        .tab({ shift: true })
+        .enterParagraphs(['\n\n1. Hey I am an ordered list in depth 0.'])
+        .enterParagraphs(['\n\n- Hey I am an unordered list in depth 1.'])
+        .tab()
+        .enterParagraphs(['\n Hey I am an unordered list in depth 2.'])
+        .tab()
+        .enterParagraphs(['\n Hey I am an unordered list in depth 1.'])
+        .tab({ shift: true })
+        .enterParagraphs(['\n\n- Hey I am an unordered list in depth 0.']);
       cy.eyesCheckWindow(this.test.title);
     });
   });