+++ e2e/tests/text.e2e.js
@@ -103,4 +103,20 @@ describe('text', () => {
       .setTextStyle(INLINE_TOOLBAR_BUTTONS.UNORDERED_LIST, [550, 1]);
     cy.eyesCheckWindow(this.test.title);
   });
+
+  it('allow to create link and open link toolbar (InlinePluginToolbar)', function() {
+    cy.loadEditorAndViewer('plain')
+      .setLink([0, 10], 'https://www.wix.com/')
+      .setSelection(5, 0)
+      .wait(200);
+    cy.eyesCheckWindow(this.test.title);
+  });
+
+  it('allow to enter tab character', function() {
+    cy.loadEditorAndViewer()
+      .focusEditor()
+      .tab()
+      .enterParagraphs(['How to eat healthy is a good question.']);
+    cy.eyesCheckWindow(this.test.title);
+  });
 });