+++ packages/editor-common/web/src/Utils/draftUtils.js
@@ -26,6 +26,15 @@ export const insertLinkInPosition = (
   });
 };
 
+export const updateLinkAtCurrentSelection = (editorState, data) => {
+  const selection = getSelection(editorState);
+  const editorStateWithLink = updateLink(selection, editorState, data);
+  return EditorState.forceSelection(
+    editorStateWithLink,
+    selection.merge({ anchorOffset: selection.focusOffset })
+  );
+};
+
 export const getBlockAtStartOfSelection = editorState => {
   const selectionState = editorState.getSelection();
   const contentState = editorState.getCurrentContent();
@@ -45,10 +54,7 @@ export const insertLinkAtCurrentSelection = (editorState, data) => {
   }
   let editorStateWithLink;
   if (isSelectionBelongsToExsistingLink(newEditorState, selection)) {
-    const blockKey = selection.getStartKey();
-    const block = newEditorState.getCurrentContent().getBlockForKey(blockKey);
-    const entityKey = block.getEntityAt(selection.getStartOffset());
-    editorStateWithLink = setEntityData(newEditorState, entityKey, createLinkEntityData(data));
+    editorStateWithLink = updateLink(selection, newEditorState, data);
   } else {
     editorStateWithLink = insertLink(newEditorState, selection, data);
   }
@@ -65,6 +71,13 @@ function isSelectionBelongsToExsistingLink(editorState, selection) {
   return getSelectedLinks(editorState).find(({ range }) => {
     return range[0] <= startOffset && range[1] >= endOffset;
   });
+}
+
+function updateLink(selection, editorState, data) {
+  const blockKey = selection.getStartKey();
+  const block = editorState.getCurrentContent().getBlockForKey(blockKey);
+  const entityKey = block.getEntityAt(selection.getStartOffset());
+  return setEntityData(editorState, entityKey, createLinkEntityData(data));
 }
 
 function preventLinkInlineStyleForNewLine(editorState, { anchorKey, focusOffset }) {
@@ -99,10 +112,12 @@ function insertLink(editorState, selection, data) {
 }
 
 function createLinkEntityData({ url, targetBlank, nofollow, anchorTarget, relValue }) {
+  const target = targetBlank ? '_blank' : anchorTarget !== '_blank' ? anchorTarget : '_self';
+  const rel = nofollow ? 'nofollow' : relValue !== 'nofollow' ? relValue : 'noopener';
   return {
     url,
-    target: targetBlank ? '_blank' : anchorTarget || '_self',
-    rel: nofollow ? 'nofollow' : relValue || 'noopener noreferrer',
+    target,
+    rel,
   };
 }
 
@@ -275,8 +290,21 @@ export const deleteBlock = (editorState, blockKey) => {
     anchorOffset,
     focusKey: blockKey,
     focusOffset: block.text.length,
+    hasFocus: true,
   });
   const newContentState = Modifier.removeRange(contentState, selectionRange, 'forward');
+  return EditorState.push(editorState, newContentState, 'remove-range');
+};
+
+export const deleteBlockText = (editorState, blockKey) => {
+  const contentState = editorState.getCurrentContent();
+  const block = contentState.getBlockForKey(blockKey);
+  const selectionRange = createSelection({
+    blockKey,
+    anchorOffset: 0,
+    focusOffset: block.text.length,
+  });
+  const newContentState = Modifier.replaceText(contentState, selectionRange, '');
   return EditorState.push(editorState, newContentState, 'remove-range');
 };
 
@@ -488,6 +516,24 @@ export function getBlockInfo(editorState, blockKey) {
   return { type: type || 'text', entityData };
 }
 
+export function getBlockType(editorState) {
+  const contentState = editorState.getCurrentContent();
+  const blockKey = editorState.getSelection().getAnchorKey();
+  const block = contentState.getBlockForKey(blockKey);
+  return block.type;
+}
+
 export function setSelection(editorState, selection) {
   return EditorState.acceptSelection(editorState, selection);
+}
+
+export function setForceSelection(editorState, selection) {
+  return EditorState.forceSelection(editorState, selection);
+}
+
+export function insertString(editorState, string) {
+  const contentState = editorState.getCurrentContent();
+  const selection = editorState.getSelection();
+  const newContentState = Modifier.replaceText(contentState, selection, string);
+  return EditorState.push(editorState, newContentState, 'insert-string');
 }