+++ packages/editor-common/web/src/index.js
@@ -15,7 +15,6 @@ export { default as BUTTONS } from './Base/buttons/keys';
 export { default as createBasePlugin } from './Base/createBasePlugin';
 export { default as createBaseComponent } from './Base/createBaseComponent';
 export { default as createBaseInsertPluginButton } from './Base/createBaseInsertPluginButton';
-export { default as createBaseToolbar } from './Base/createBaseToolbar';
 export { default as baseToolbarButton } from './Base/baseToolbarButton';
 
 // Components
@@ -59,6 +58,7 @@ export { simplePubsub } from './Utils/simplePubsub';
 export { getModalStyles, getBottomToolbarModalStyles } from './Utils/getModalStyles';
 
 export {
+  updateLinkAtCurrentSelection,
   insertLinkAtCurrentSelection,
   insertLinkInPosition,
   hasLinksInBlock,
@@ -85,10 +85,13 @@ export {
   calculateDiff,
   getPostContentSummary,
   createSelection,
+  getBlockType,
+  setForceSelection,
+  deleteBlockText,
+  insertString,
 } from './Utils/draftUtils';
 export { isiOS } from './Utils/isiOS';
 export { getSelectionStyles } from './Utils/inlineStyleUtils';
-export { getConfigByFormFactor } from './Utils/getConfigByFormFactor';
 export { mergeToolbarSettings } from './Utils/mergeToolbarSettings';
 export {
   COMMANDS,
@@ -99,6 +102,8 @@ export {
   DECORATION_MODE,
   PLUGIN_DECORATION_PROPS,
   PLUGIN_DECORATIONS,
+  TEXT_TYPES,
+  CHARACTERS,
 } from './consts';
 
 export {