+++ packages/editor/web/src/RichContentEditor/RichContentEditor.jsx
@@ -22,6 +22,9 @@ import {
   calculateDiff,
   getPostContentSummary,
   Modifier,
+  getBlockType,
+  COMMANDS,
+  MODIFIERS,
 } from 'wix-rich-content-editor-common';
 
 import {
@@ -32,8 +35,9 @@ import {
 } from 'wix-rich-content-common';
 import styles from '../../statics/styles/rich-content-editor.scss';
 import draftStyles from '../../statics/styles/draft.rtlignore.scss';
+import 'wix-rich-content-common/dist/statics/styles/draftDefault.rtlignore.scss';
 import { convertFromHTML as draftConvertFromHtml } from 'draft-convert';
-import pastedContentConfig from './utils/pastedContentConfig';
+import { pastedContentConfig, clearUnnecessaryInlineStyles } from './utils/pastedContentUtil';
 
 class RichContentEditor extends Component {
   static getDerivedStateFromError(error) {
@@ -219,30 +223,36 @@ class RichContentEditor extends Component {
     if (handlePastedText) {
       return handlePastedText(text, html, editorState);
     }
-    let contentState;
     if (html) {
       const htmlContentState = draftConvertFromHtml(pastedContentConfig)(html);
-      contentState = Modifier.replaceWithFragment(
+      const contentState = Modifier.replaceWithFragment(
         editorState.getCurrentContent(),
         editorState.getSelection(),
         htmlContentState.getBlockMap()
       );
+      const newEditorState = EditorState.push(editorState, contentState, 'insert-fragment');
+      const newContentState = clearUnnecessaryInlineStyles(contentState);
+      const resultEditorState = EditorState.set(newEditorState, {
+        currentContent: newContentState,
+        selection: newEditorState.getSelection(),
+      });
+
+      this.updateEditorState(resultEditorState);
+      return 'handled';
     } else {
-      contentState = Modifier.replaceText(
-        editorState.getCurrentContent(),
-        editorState.getSelection(),
-        text
-      );
+      return false;
     }
+  };
 
-    const newEditorState = EditorState.push(editorState, contentState, 'insert-fragment');
-    const resultEditorState = EditorState.set(newEditorState, {
-      currentContent: contentState,
-      selection: newEditorState.getSelection(),
-    });
-
-    this.updateEditorState(resultEditorState);
-    return 'handled';
+  handleTabCommand = () => {
+    if (this.getToolbars().TextToolbar) {
+      const staticToolbarButton = this.findFocusableChildForElement(
+        `${getStaticTextToolbarId(this.refId)}`
+      );
+      staticToolbarButton && staticToolbarButton.focus();
+    } else {
+      this.editor.blur();
+    }
   };
 
   handlePastedText = (text, html, editorState) => {
@@ -272,23 +282,20 @@ class RichContentEditor extends Component {
     commands: [
       ...this.pluginKeyBindings.commands,
       {
-        command: 'tab',
+        command: COMMANDS.TAB,
         modifiers: [],
+        key: 'Tab',
+      },
+      {
+        command: COMMANDS.SHIFT_TAB,
+        modifiers: [MODIFIERS.SHIFT],
         key: 'Tab',
       },
     ],
     commandHanders: {
       ...this.pluginKeyBindings.commandHandlers,
-      tab: () => {
-        if (this.getToolbars().TextToolbar) {
-          const staticToolbarButton = this.findFocusableChildForElement(
-            `${getStaticTextToolbarId(this.refId)}`
-          );
-          staticToolbarButton && staticToolbarButton.focus();
-        } else {
-          this.editor.blur();
-        }
-      },
+      tab: this.handleTabCommand,
+      shiftTab: this.handleTabCommand,
     },
   });
 
@@ -327,7 +334,8 @@ class RichContentEditor extends Component {
     ];
     //eslint-disable-next-line array-callback-return
     const toolbars = this.plugins.map((plugin, index) => {
-      const Toolbar = plugin.Toolbar || plugin.InlineToolbar || plugin.SideToolbar;
+      const Toolbar =
+        plugin.Toolbar || plugin.InlinePluginToolbar || plugin.InlineToolbar || plugin.SideToolbar;
       if (Toolbar) {
         if (includes(toolbarsToIgnore, plugin.name)) {
           return null;
@@ -393,7 +401,8 @@ class RichContentEditor extends Component {
         blockStyleFn={blockStyleFn(theme, this.styleToClass)}
         handleKeyCommand={handleKeyCommand(
           this.updateEditorState,
-          this.getCustomCommandHandlers().commandHanders
+          this.getCustomCommandHandlers().commandHanders,
+          getBlockType(editorState)
         )}
         editorKey={editorKey}
         keyBindingFn={createKeyBindingFn(this.getCustomCommandHandlers().commands || [])}