+++ packages/plugin-file-upload/web/src/viewer.js
@@ -1,12 +1,13 @@
 import { typeMapper } from './typeMapper';
 import { FILE_UPLOAD_TYPE } from './types';
 export { typeMapper as fileUploadTypeMapper, FILE_UPLOAD_TYPE };
-import { DEFAULTS } from './consts';
+import { DEFAULTS, THEME as theme } from './consts';
 
 export const pluginFileUpload = (config = {}) => {
   return {
     config: { ...DEFAULTS.configViewer, ...config },
     type: FILE_UPLOAD_TYPE,
     typeMapper,
+    theme,
   };
 };