+++ packages/plugin-link-preview/web/src/editor.js
@@ -1,6 +1,6 @@
 import { createLinkPreviewPlugin } from './createLinkPreviewPlugin';
 import { LINK_PREVIEW_TYPE } from './types';
-import { DEFAULTS } from './defaults';
+import { DEFAULTS, THEME as theme } from './defaults';
 
 export const pluginLinkPreview = (config = {}) => {
   return {
@@ -8,5 +8,6 @@ export const pluginLinkPreview = (config = {}) => {
     type: LINK_PREVIEW_TYPE,
     createPlugin: createLinkPreviewPlugin,
     ModalsMap: {},
+    theme,
   };
 };