+++ packages/preview/web/src/Components/ImageCounter.jsx
@@ -2,6 +2,8 @@ import React, { PureComponent } from 'react';
 import { mergeStyles } from 'wix-rich-content-common';
 import ReactDOM from 'react-dom';
 import PropTypes from 'prop-types';
+import { debounce } from 'lodash';
+import Measure from 'react-measure';
 import styles from '../../statics/styles/image-counter.scss';
 
 class ImageCounter extends PureComponent {
@@ -45,20 +47,20 @@ class ImageCounter extends PureComponent {
     );
   };
 
-  decorateImages() {
-    setTimeout(() => {
-      if (this.wrapper) {
-        const images = this.wrapper.querySelectorAll('[role=img]');
-        const imagesToDecorate = this.props.imageSelector(images);
-        const decorations = imagesToDecorate.map(img => this.renderDecoration(img));
-        ReactDOM.render(decorations, this.container);
-      }
-    }, 500);
-  }
+  decorateImages = () => {
+    if (this.wrapper) {
+      const images = this.wrapper.querySelectorAll('[role=img]');
+      const imagesToDecorate = this.props.imageSelector(images);
+      const decorations = imagesToDecorate.map(img => this.renderDecoration(img));
+      ReactDOM.render(decorations, this.container);
+    }
+  };
 
   componentDidMount() {
     this.decorateImages();
   }
+
+  onResize = debounce(this.decorateImages, 200);
 
   handleWrapper = el => (this.wrapper = el);
 
@@ -68,10 +70,16 @@ class ImageCounter extends PureComponent {
     this.styles = this.styles || mergeStyles({ styles, theme: this.props.theme });
     /* eslint-disable */
     return (
-      <div ref={this.handleWrapper} onClick={this.onClick}>
-        <div ref={this.handleContainer} className={this.styles.imageCounter_overlay} />
-        {this.props.children}
-      </div>
+      <Measure onResize={this.onResize}>
+        {({ measureRef }) => (
+          <div ref={measureRef}>
+            <div ref={this.handleWrapper} onClick={this.onClick}>
+              <div ref={this.handleContainer} className={this.styles.imageCounter_overlay} />
+              {this.props.children}
+            </div>
+          </div>
+        )}
+      </Measure>
     );
     /* eslint-enable */
   }