+++ packages/preview/web/src/RuleEngine/ContentStateTransformation.spec.js
@@ -10,14 +10,17 @@ describe('Content State Transformation', () => {
         {
           metadata: { height: 1920, width: 1920 },
           url: '8bb438_c1089eafb4ab405ba328b528e3ecc63e.jpg',
+          itemId: '8bb438_c1089eafb4ab405ba328b528e3ecc63e.jpg',
         },
         {
           metadata: { height: 2800, width: 4200 },
           url: '8bb438_e78b371c75ce42de8719dccfc97298a4.jpg',
+          itemId: '8bb438_e78b371c75ce42de8719dccfc97298a4.jpg',
         },
         {
           metadata: { height: 1280, width: 1920 },
           url: '8bb438_281af3d3281f4584a5a864c6c60f3a00.jpg',
+          itemId: '8bb438_281af3d3281f4584a5a864c6c60f3a00.jpg',
         },
       ],
       styles: {