+++ packages/viewer/web/src/RichContentViewer.jsx
@@ -7,6 +7,7 @@ import {
   normalizeInitialState,
   getLangDir,
 } from 'wix-rich-content-common';
+import 'wix-rich-content-common/dist/statics/styles/draftDefault.rtlignore.scss';
 import { convertToReact } from './utils/convertContentState';
 import viewerStyles from '../statics/rich-content-viewer.scss';
 import viewerAlignmentStyles from '../statics/rich-content-viewer-alignment.rtlignore.scss';